feat: IssueDataInterface type 분리 공통되는 IssueDataInterface type을 분리
