+++ src/contexts/api.tsx
@@ -1,16 +1,5 @@
 import React, { createContext, Dispatch, useContext, useReducer } from 'react';
-
-export interface IssueDataInterface {
-  title: string;
-  body: string;
-  number: number;
-  created_at: string;
-  user: {
-    login: string;
-    avatar_url: string;
-  };
-  [key: string]: any;
-}
+import { IssueDataInterface } from '../types/type';
 
 type State = {
   issueList: {
